+++ src/lib.rs
@@ -17,13 +17,14 @@ use std::io::{Error, ErrorKind};
 pub fn run() -> Result<(), Box<dyn std::error::Error>> {
     match cli::from_args() {
         Command::DockerInfo => docker::docker_version()?,
-        Command::Wifi(WifiSet { config, image }) => docker::set_wifi_config(&config, &image)?,
+        Command::Wifi(WifiSet { config, image, generate_bmap }) => docker::set_wifi_config(&config, &image, generate_bmap)?,
         Command::Enrollment(EnrollmentSet {
             enrollment_config,
             image,
-        }) => docker::set_enrollment_config(&enrollment_config, &image)?,
-        Command::Identity(SetConfig { config, image }) => {
-            docker::set_identity_config(&config, &image)?
+            generate_bmap,
+        }) => docker::set_enrollment_config(&enrollment_config, &image, generate_bmap)?,
+        Command::Identity(SetConfig { config, image, generate_bmap }) => {
+            docker::set_identity_config(&config, &image, generate_bmap)?
         }
         Command::Identity(SetIotedgeGatewayConfig {
             config,
@@ -31,22 +32,26 @@ pub fn run() -> Result<(), Box<dyn std::error::Error>> {
             root_ca,
             device_identity,
             device_identity_key,
+            generate_bmap,
         }) => docker::set_iotedge_gateway_config(
             &config,
             &image,
             &root_ca,
             &device_identity,
             &device_identity_key,
+            generate_bmap,
         )?,
         Command::Identity(SetIotLeafSasConfig {
             config,
             image,
             root_ca,
-        }) => docker::set_iot_leaf_sas_config(&config, &image, &root_ca)?,
+            generate_bmap,
+        }) => docker::set_iot_leaf_sas_config(&config, &image, &root_ca, generate_bmap)?,
         Command::IotHubDeviceUpdate(IotHubDeviceUpdateSet {
             iot_hub_device_update_config,
             image,
-        }) => docker::set_iot_hub_device_update_config(&iot_hub_device_update_config, &image)?,
+            generate_bmap
+        }) => docker::set_iot_hub_device_update_config(&iot_hub_device_update_config, &image, generate_bmap)?,
         _ => Err(Error::new(ErrorKind::Other, "Not implemented"))?,
     }