+++ tests/integration_tests.rs
@@ -7,14 +7,27 @@ extern crate lazy_static;
 
 #[test]
 fn check_set_wifi_config() {
-    let tr = Testrunner::new("check_set_wifi_config");
+    let tr = Testrunner::new(function_name!().split("::").last().unwrap());
 
     let config_file_path = tr.to_pathbuf("wpa_supplicant.conf");
     let image_path = tr.to_pathbuf("image.wic");
 
     assert_eq!(
         true,
-        docker::set_wifi_config(&config_file_path, &image_path).is_ok()
+        docker::set_wifi_config(&config_file_path, &image_path, false).is_ok()
+    );
+}
+
+//#[test]
+fn check_set_wifi_config_bmap() {
+    let tr = Testrunner::new(function_name!().split("::").last().unwrap());
+
+    let config_file_path = tr.to_pathbuf("wpa_supplicant.conf");
+    let image_path = tr.to_pathbuf("image.wic");
+
+    assert_eq!(
+        true,
+        docker::set_wifi_config(&config_file_path, &image_path, true).is_ok()
     );
 }
 
@@ -27,7 +40,20 @@ fn check_set_enrollment_config() {
 
     assert_eq!(
         true,
-        docker::set_enrollment_config(&enrollment_config_file_path, &image_path).is_ok()
+        docker::set_enrollment_config(&enrollment_config_file_path, &image_path, false).is_ok()
+    );
+}
+
+//#[test]
+fn check_set_enrollment_config_bmap() {
+    let tr = Testrunner::new(function_name!().split("::").last().unwrap());
+
+    let enrollment_config_file_path = tr.to_pathbuf("enrollment_static.json");
+    let image_path = tr.to_pathbuf("image.wic");
+
+    assert_eq!(
+        true,
+        docker::set_enrollment_config(&enrollment_config_file_path, &image_path, true).is_ok()
     );
 }
 
@@ -40,7 +66,7 @@ fn check_set_enrollment_config_missing_dps() {
 
     assert_ne!(
         None,
-        docker::set_enrollment_config(&enrollment_config_file_path, &image_path)
+        docker::set_enrollment_config(&enrollment_config_file_path, &image_path, false)
             .unwrap_err()
             .to_string()
             .find("missing field `dpsConnectionString`")
@@ -56,7 +82,7 @@ fn check_set_enrollment_config_missing_iothub() {
 
     assert_ne!(
         None,
-        docker::set_enrollment_config(&enrollment_config_file_path, &image_path)
+        docker::set_enrollment_config(&enrollment_config_file_path, &image_path, false)
             .unwrap_err()
             .to_string()
             .find("missing field `iothubConnectionString`")
@@ -72,7 +98,7 @@ fn check_set_enrollment_config_unknown_key() {
 
     assert_ne!(
         None,
-        docker::set_enrollment_config(&enrollment_config_file_path, &image_path)
+        docker::set_enrollment_config(&enrollment_config_file_path, &image_path, false)
             .unwrap_err()
             .to_string()
             .find("unknown field `someKeyWeDontKnow`")
@@ -89,7 +115,7 @@ fn check_set_enrollment_config_invalid_connection_string() {
 
     assert_ne!(
         None,
-        docker::set_enrollment_config(&enrollment_config_file_path, &image_path)
+        docker::set_enrollment_config(&enrollment_config_file_path, &image_path, false)
             .unwrap_err()
             .to_string()
             .find("Enrollment validation failed")
@@ -98,7 +124,7 @@ fn check_set_enrollment_config_invalid_connection_string() {
 
 #[test]
 fn check_set_identity_gateway_config() {
-    let tr = Testrunner::new("check_set_identity_gateway_config");
+    let tr = Testrunner::new(function_name!().split("::").last().unwrap());
 
     let config_file_path = tr.to_pathbuf("identity_config_gateway.toml");
     let image_path = tr.to_pathbuf("image.wic");
@@ -113,7 +139,32 @@ fn check_set_identity_gateway_config() {
             &image_path,
             &root_ca_file_path,
             &edge_device_identity_full_chain_file_path,
-            &edge_device_identity_key_file_path
+            &edge_device_identity_key_file_path,
+            false
+        )
+        .is_ok()
+    );
+}
+
+//#[test]
+fn check_set_identity_gateway_config_bmap() {
+    let tr = Testrunner::new(function_name!().split("::").last().unwrap());
+
+    let config_file_path = tr.to_pathbuf("identity_config_gateway.toml");
+    let image_path = tr.to_pathbuf("image.wic");
+    let root_ca_file_path = tr.to_pathbuf("root.ca.cert.pem");
+    let edge_device_identity_full_chain_file_path = tr.to_pathbuf("full-chain.cert.pem");
+    let edge_device_identity_key_file_path = tr.to_pathbuf("device-ca.key.pem");
+
+    assert_eq!(
+        true,
+        docker::set_iotedge_gateway_config(
+            &config_file_path,
+            &image_path,
+            &root_ca_file_path,
+            &edge_device_identity_full_chain_file_path,
+            &edge_device_identity_key_file_path,
+            true
         )
         .is_ok()
     );
@@ -121,7 +172,7 @@ fn check_set_identity_gateway_config() {
 
 #[test]
 fn check_set_identity_leaf_config() {
-    let tr = Testrunner::new("check_set_identity_leaf_config");
+    let tr = Testrunner::new(function_name!().split("::").last().unwrap());
 
     let config_file_path = tr.to_pathbuf("identity_config_leaf.toml");
     let image_path = tr.to_pathbuf("image.wic");
@@ -129,32 +180,73 @@ fn check_set_identity_leaf_config() {
 
     assert_eq!(
         true,
-        docker::set_iot_leaf_sas_config(&config_file_path, &image_path, &root_ca_file_path).is_ok()
+        docker::set_iot_leaf_sas_config(&config_file_path, &image_path, &root_ca_file_path, false).is_ok()
+    );
+}
+
+//#[test]
+fn check_set_identity_leaf_config_bmap() {
+    let tr = Testrunner::new(function_name!().split("::").last().unwrap());
+
+    let config_file_path = tr.to_pathbuf("identity_config_leaf.toml");
+    let image_path = tr.to_pathbuf("image.wic");
+    let root_ca_file_path = tr.to_pathbuf("root.ca.cert.pem");
+
+    assert_eq!(
+        true,
+        docker::set_iot_leaf_sas_config(&config_file_path, &image_path, &root_ca_file_path, true).is_ok()
     );
 }
 
 #[test]
 fn check_set_identity_config() {
-    let tr = Testrunner::new("check_set_identity_config");
+    let tr = Testrunner::new(function_name!().split("::").last().unwrap());
 
     let config_file_path = tr.to_pathbuf("identity_config_dps_tpm.toml");
     let image_path = tr.to_pathbuf("image.wic");
 
     assert_eq!(
         true,
-        docker::set_identity_config(&config_file_path, &image_path).is_ok()
+        docker::set_identity_config(&config_file_path, &image_path, false).is_ok()
+    );
+}
+
+//#[test]
+fn check_set_identity_config_bmap() {
+    let tr = Testrunner::new(function_name!().split("::").last().unwrap());
+
+    let config_file_path = tr.to_pathbuf("identity_config_dps_tpm.toml");
+    let image_path = tr.to_pathbuf("image.wic");
+
+    assert_eq!(
+        true,
+        docker::set_identity_config(&config_file_path, &image_path, true).is_ok()
     );
 }
 
 #[test]
 fn check_set_iot_hub_device_update_config() {
-    let tr = Testrunner::new(&"check_set_iot_hub_device_update_config");
+    let tr = Testrunner::new(function_name!().split("::").last().unwrap());
 
     let adu_config_file_path = tr.to_pathbuf("adu-conf.txt");
     let image_path = tr.to_pathbuf("image.wic");
 
     assert_eq!(
         true,
-        docker::set_iot_hub_device_update_config(&adu_config_file_path, &image_path).is_ok()
+        docker::set_iot_hub_device_update_config(&adu_config_file_path, &image_path, false).is_ok()
+    );
+}
+
+
+//#[test]
+fn check_set_iot_hub_device_update_config_bmap() {
+    let tr = Testrunner::new(function_name!().split("::").last().unwrap());
+
+    let adu_config_file_path = tr.to_pathbuf("adu-conf.txt");
+    let image_path = tr.to_pathbuf("image.wic");
+
+    assert_eq!(
+        true,
+        docker::set_iot_hub_device_update_config(&adu_config_file_path, &image_path, true).is_ok()
     );
 }