+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.0] Q4 2021
+- optionally create bmap file for commands 'enrollment', 'identity',
+  'iot-hub-device-update' and 'wiki'
+
 ## [0.7.0] Q4 2021
 
 - rewrite backend scripts to use e2tool, so we don't need privileged docker context